Finish Game Logic and moves
This is the second version of the chess game I wrote.
Due to the shitty data structure and programming design,
further extension to the program is complicated. The overall
design needs to be refactored. The architecture of the program
needs thorough consideration and planning beforehand. Lack of
experience and programming skills leads to this messy code.
It's not an easy job. This project probably ends here.